requirements/base.txt
django_tastypie>=0.9.16
django-oauth-toolkit>=0.7.2
python-mimeparse>=0.1.4
# or
# django-oauth2-provider>=0.2.6.1
